Do not use compressed ciphertexts
The method to load compressed ciphertexts from a stream is not working. Apparently because of the latest changes to SEAL. Before, when a Ciphertext was constructed it started with a size of 2. But now it starts with a size of 0, and the compressed_load method in CiphertextCompressor has now way to initialize the size of the Ciphertext.
For now reverting to not using compressed ciphertexts between the sender and receiver.