Allow setting the Sender DB again
Setting the DB again was crashing because the plaintexts in batch_random_symm_poly_storage_ are left in NTT form after
being preprocessed, and loading the DB performs operations that are
invalid for NTT form plaintexts.
I also had to clear the offline compute progress earlier, as the result of the previous load lingers in memory aparently.